requirements-backend.txt
sqlalchemy==1.4.36
asyncio==3.4.3
alembic==1.7.7
psycopg2==2.8.4
pydantic-sqlalchemy==0.0.9
asyncpg==0.23.0
fastapi==0.68.0
uvicorn==0.14.0
